visualization/requirements.txt
argparse
matplotlib
numpy
open3d
pyyaml